backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Ollama client
ollama==0.1.6

# Vector DB
pinecone-client==3.0.0
sentence-transformers==2.7.0
huggingface-hub==0.20.0

# PDF parsing
PyPDF2==3.0.1

# Utilities
aiofiles==23.2.1